+++ src/projects/constants.py
@@ -28,9 +28,11 @@ HOME_DIR = str(Path.home())
 CONFIG_PATH = Path(user_config_dir(__app_name__, __author__), "config.toml")
 USER_DATA_DIR = Path(user_data_dir(__app_name__, __author__))
 USER_DATA_DIR.mkdir(exist_ok=True)
+
 ICON_FILE = Path(Path(__file__).parent, "images", "rocket-launch-outline.png")
 ICON_DIR = f"{Path(__file__).parent}/icons/"
 USER_DATA_FILE = Path(USER_DATA_DIR, "data.json")
+CACHED_ENVS_FILE = "cached_envs.json"
 
 # GUI
 APP_TITLE = "Project management"