+++ src/projects/project_server.py
@@ -2,13 +2,10 @@
 
 from pathlib import Path
 
-import json5
-
 import projects.projects_io as io
 from projects.config import config
-from projects.constants import HOME_DIR, USER_DATA_DIR
-from projects.env_version import EnvironmentVersion
-from projects.project import DEFAULT_COLOURS, Project
+from projects.constants import CACHED_ENVS_FILE, USER_DATA_DIR
+from projects.project import Project
 
 
 class ProjectServer:
@@ -21,29 +18,14 @@ class ProjectServer:
     def _get_projects(self) -> dict[str, Project]:
         project_dict = {}
         projects_raw = io.read_json_file(self.project_file)
+        cached_envs = self._get_cached_envs()
         for key, item in projects_raw.items():
-            project = Project()
-            project.name = key
-            project_dict[key] = project
+            item["name"] = key
+            item["cached_envs"] = cached_envs.get(key, {})
 
-            project.source_dir = item["source_dir"].replace("~", HOME_DIR)
-            project.base_dir = item["base_dir"].replace("~", HOME_DIR)
-            project.pypi = item["pypi"]
-            if "build_for_windows" not in item:
-                item["build_for_windows"] = False
-            project.build_for_windows = item["build_for_windows"]
-            if "repository" in item:
-                project.repository_name = item["repository"]
-            project.cached_envs = {
-                key: EnvironmentVersion(data)
-                for key, data in item["cached_envs"].items()
-            }
-            if "script" in item:
-                project.script = item["script"]
-            if "desktop_file" in item:
-                project.desktop_file = item["desktop_file"]
-            project.get_project_data()
-            self.get_project_colours(project)
+            project = Project()
+            project.deserialize(item)
+            project_dict[key] = project
         return project_dict
 
     def save_projects(self, projects: dict[str, Project] = None) -> int:
@@ -53,44 +35,34 @@ class ProjectServer:
         output = {
             name: project.serialize() for name, project in projects.items()
         }
+
+        cached_envs = {}
         for project in self.projects.values():
-            self.save_project_colours(project)
+            project.save_project_colours()
+            cached_envs[project.name] = {
+                key: item for key, item in project.cached_envs.items()
+            }
+        if cached_envs:
+            self._save_cached_envs(cached_envs)
         return io.update_json_file(self.project_file, output)
 
-    def get_project_colours(self, project: Project) -> None:
-        """Get the colours for a project."""
-
-        settings_file = Path(
-            Path(project.source_dir).parent.parent, ".vscode", "settings.json"
-        )
+    def _get_cached_envs(self) -> dict:
+        """Get the cached environments for a project."""
+        cached_envs_file = Path(USER_DATA_DIR, CACHED_ENVS_FILE)
         try:
-            with open(settings_file, encoding="utf-8") as f:
-                settings = json5.load(f)
+            cached_envs = io.read_json_file(cached_envs_file)
+            return cached_envs
         except FileNotFoundError:
-            return
+            return {}
 
-        if "workbench.colorCustomizations" in settings:
-            colour_customizations = settings["workbench.colorCustomizations"]
-            for key, value in colour_customizations.items():
-                if key in project.workbench_colours:
-                    project.workbench_colours[key] = value
-
-    def save_project_colours(self, project: Project) -> None:
-        """Save the colours for a project."""
-        settings_file = Path(
-            Path(project.source_dir).parent.parent, ".vscode", "settings.json"
-        )
-        try:
-            with open(settings_file, encoding="utf-8") as f:
-                settings = json5.load(f)
-        except FileNotFoundError:
-            return
-        if project.workbench_colours != DEFAULT_COLOURS:
-            settings["workbench.colorCustomizations"] = (
-                project.workbench_colours
-            )
-            with open(settings_file, "w", encoding="utf-8") as f:
-                json5.dump(settings, f, indent=2)
-            #     print(project.name)
-            # print(settings[
-            #     "workbench.colorCustomizations"])
+    def _save_cached_envs(self, cached_envs: dict) -> None:
+        output = {}
+        for project_name, env in cached_envs.items():
+            project_envs = {}
+            for key, item in env.items():
+                project_envs[key] = item.serialize()
+            if project_envs:
+                output[project_name] = project_envs
+        """Save the cached environments for a project."""
+        cached_envs_file = Path(USER_DATA_DIR, CACHED_ENVS_FILE)
+        io.update_json_file(cached_envs_file, output)