+++ src/projects/text.py
@@ -14,6 +14,8 @@ from dataclasses import dataclass, field
 from psiutils.text import Text as PsiText
 
 strings = {
+    "ABOUT": "About",
+    "AUTHOR": "Author",
     "ACCEPT": "Accept",
     "BUILD_FOR_WINDOWS": "Build for Windows",
     "CONFIG": "Preferences",
@@ -26,8 +28,6 @@ strings = {
     "RUN_SCRIPT": "Run script",
     "SELECT": "Select",
     "WINDSURF": "Windsurf",
-    "ABOUT": "About",
-    "AUTHOR": "Author",
     "DATA_DIRECTORY": "Data directory",
     "ONLINE_HELP": "Online help",
     "VERSION": " Version",